+++ idea/testData/codeInsight/generate/toString/singleTemplate/customAccessors.kt
@@ -1,0 +1,8 @@
+class Test {
+    var serial: String = ""
+        set(value) {
+            field = value.toUpperCase()
+        }
+    var name: String = ""
+    <caret>
+}